+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.11] - 2024-01-16
+
+### Fixed
+- Do not delete a patch if it is empty.
+
+### Added
+- Validating for generated patch.
+
+## Changed
+- Write generated patch to folder with new filter.
+
+
 ## [1.0.10] - 2024-01-11
 
 ### Fixed